requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase==2.6.0
pydantic==2.8.2
httpx==0.27.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
apscheduler>=3.10.4
pytz>=2023.3